some more tests
